ROSAENG-15779: most of the content of the cmd/rhobs/requests.go has been moved in the other files of the package in order to reduce the file size
