+++ cmd/rhobs/hcpDashboard.go
@@ -2,9 +2,12 @@ package rhobs
 
 import (
 	"fmt"
+	"net/url"
 	"strings"
 
+	ocmutils "github.com/openshift/osdctl/pkg/utils"
 	"github.com/pkg/browser"
+	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
@@ -96,4 +99,250 @@ func newCmdHcpDashboard() *cobra.Command {
 	cmd.Flags().BoolVarP(&isOpeningGrafanaUrl, "browser", "b", false, "Open in the URL in the default browser")
 
 	return cmd
+}
+
+type GrafanaDashboard struct {
+	name                 string
+	pathId               string
+	pathName             string
+	validateAndGetParams func(metricsFetcher, logsFetcher *RhobsFetcher) (url.Values, error)
+}
+
+const defaultGrafanaDashboardShortName = "hosted-cluster"
+
+func eventuallyWarnAboutWiderDashboardScope(fetcher *RhobsFetcher) {
+	if fetcher.clusterId != "" {
+		log.Warnf("Dashboard will contain data for clusters other than the provided cluster. "+
+			"Works as if the --rhobs-cell option was set to the %s cluster RHOBS cell(s).\n", fetcher.clusterId)
+	}
+}
+
+var allowedGrafanaDashboard = []*GrafanaDashboard{
+	{
+		name:     defaultGrafanaDashboardShortName,
+		pathId:   "cf6ntunq7rb40c",
+		pathName: "rosa-hcp-central-cluster-dashboard",
+		validateAndGetParams: func(metricsFetcher, logsFetcher *RhobsFetcher) (url.Values, error) {
+			if !metricsFetcher.IsHostedCluster {
+				return nil, fmt.Errorf("'%s' dashboard must be used with a hosted cluster", defaultGrafanaDashboardShortName)
+			}
+
+			region, shard, err := metricsFetcher.getRhobsRegionAndShard()
+			if err != nil {
+				return nil, err
+			}
+
+			return url.Values{
+				"var-environment": {metricsFetcher.ocmEnvName},
+				"var-region":      {region},
+				"var-shard":       {shard},
+				"var-_id":         {metricsFetcher.clusterExternalId},
+			}, nil
+		},
+	}, {
+		name:     "management-cluster",
+		pathId:   "rosa-hcp-mc-dashboard",
+		pathName: "rosa-hcp-management-cluster-dashboard",
+		validateAndGetParams: func(metricsFetcher, logsFetcher *RhobsFetcher) (url.Values, error) {
+			mcName := metricsFetcher.clusterName
+
+			if !metricsFetcher.isManagementCluster {
+				if metricsFetcher.IsHostedCluster {
+					managementCluster, err := ocmutils.GetManagementCluster(metricsFetcher.clusterId)
+					if err != nil {
+						return nil, fmt.Errorf("failed to retrieve management cluster for cluster '%s': %v", metricsFetcher.clusterId, err)
+					}
+					mcName = managementCluster.Name()
+					log.Warnf("Dashboard will contain data for the %s management cluster; not just for the provided hosted cluster: %s", managementCluster.ID(), metricsFetcher.clusterId)
+				} else {
+					return nil, fmt.Errorf("'%s' dashboard must be used with a management cluster", "management-cluster")
+				}
+			}
+
+			region, _, err := metricsFetcher.getRhobsRegionAndShard()
+			if err != nil {
+				return nil, err
+			}
+
+			return url.Values{
+				"var-environment": {metricsFetcher.ocmEnvName},
+				"var-region":      {region},
+				"var-mc_name":     {mcName},
+			}, nil
+		},
+	}, {
+		name:     "kube-apis-slo",
+		pathId:   "cfmgzo0gsak1sd",
+		pathName: "drill-down3a-rosa-hcp-api-server-availability",
+		validateAndGetParams: func(metricsFetcher, logsFetcher *RhobsFetcher) (url.Values, error) {
+			eventuallyWarnAboutWiderDashboardScope(metricsFetcher)
+
+			metricsDataSource, err := metricsFetcher.getMetricsGrafanaDataSource()
+			if err != nil {
+				return nil, err
+			}
+			return url.Values{
+				"var-environment":       {metricsFetcher.ocmEnvName},
+				"var-datasource_global": {metricsDataSource},
+			}, nil
+		},
+	}, {
+		name:     "clusters-creation-slo",
+		pathId:   "fdmk9z8ucodtsa",
+		pathName: "drill-down3a-rosa-hcp-cluster-creation",
+		validateAndGetParams: func(metricsFetcher, logsFetcher *RhobsFetcher) (url.Values, error) {
+			eventuallyWarnAboutWiderDashboardScope(metricsFetcher)
+
+			region, _, err := metricsFetcher.getRhobsRegionAndShard()
+			if err != nil {
+				return nil, err
+			}
+
+			metricsDataSource, err := metricsFetcher.getMetricsGrafanaDataSource()
+			if err != nil {
+				return nil, err
+			}
+
+			logsDataSource, err := logsFetcher.getLogsGrafanaDataSource()
+			if err != nil {
+				return nil, err
+			}
+
+			return url.Values{
+				"var-environment":         {metricsFetcher.ocmEnvName},
+				"var-region":              {region},
+				"var-datasource_regional": {metricsDataSource},
+				"var-datasource_global":   {metricsDataSource},
+				"var-datasource_logs":     {logsDataSource},
+			}, nil
+		},
+	}, {
+		name:     "control-planes-upgrade-slo",
+		pathId:   "efmgzo0i3qmm8d",
+		pathName: "drill-down3a-rosa-hcp-control-plane-upgrades",
+		validateAndGetParams: func(metricsFetcher, logsFetcher *RhobsFetcher) (url.Values, error) {
+			region, _, err := metricsFetcher.getRhobsRegionAndShard()
+			if err != nil {
+				return nil, err
+			}
+
+			metricsDataSource, err := metricsFetcher.getMetricsGrafanaDataSource()
+			if err != nil {
+				return nil, err
+			}
+
+			clusterId := metricsFetcher.clusterId
+			if !metricsFetcher.IsHostedCluster {
+				clusterId = "$__all"
+			}
+
+			return url.Values{
+				"var-environment":         {metricsFetcher.ocmEnvName},
+				"var-region":              {region},
+				"var-datasource_regional": {metricsDataSource},
+				"var-datasource_global":   {metricsDataSource},
+				"var-clusterid":           {clusterId},
+			}, nil
+		},
+	}, {
+		name:     "nodepools-upgrade-slo",
+		pathId:   "919c6ec2b6d74bdf",
+		pathName: "drill-down3a-rosa-hcp-nodepool-upgrades",
+		validateAndGetParams: func(metricsFetcher, logsFetcher *RhobsFetcher) (url.Values, error) {
+			metricsDataSource, err := metricsFetcher.getMetricsGrafanaDataSource()
+			if err != nil {
+				return nil, err
+			}
+
+			clusterId := metricsFetcher.clusterId
+			if !metricsFetcher.IsHostedCluster {
+				clusterId = "$__all"
+			}
+			mcName := metricsFetcher.clusterName
+			if !metricsFetcher.isManagementCluster {
+				mcName = "$__all"
+			}
+
+			return url.Values{
+				"var-datasource":        {metricsDataSource},
+				"var-namespace":         {"uhc-" + metricsFetcher.ocmEnvName},
+				"var-clusterid":         {clusterId},
+				"var-managementcluster": {mcName},
+			}, nil
+		},
+	}, {
+		name:     "nodepools-slo",
+		pathId:   "cdtg6ugw1a03ka",
+		pathName: "drill-down3a-rosa-hcp-nodepools",
+		validateAndGetParams: func(metricsFetcher, logsFetcher *RhobsFetcher) (url.Values, error) {
+			eventuallyWarnAboutWiderDashboardScope(metricsFetcher)
+
+			region, _, err := metricsFetcher.getRhobsRegionAndShard()
+			if err != nil {
+				return nil, err
+			}
+
+			metricsDataSource, err := metricsFetcher.getMetricsGrafanaDataSource()
+			if err != nil {
+				return nil, err
+			}
+
+			return url.Values{
+				"var-environment":         {metricsFetcher.ocmEnvName},
+				"var-region":              {region},
+				"var-datasource_regional": {metricsDataSource},
+			}, nil
+		},
+	}, {
+		name:     "counters",
+		pathId:   "bfmgzo0f6uw3kc",
+		pathName: "rosa-hcp-counter",
+		validateAndGetParams: func(metricsFetcher, logsFetcher *RhobsFetcher) (url.Values, error) {
+			eventuallyWarnAboutWiderDashboardScope(metricsFetcher)
+
+			region, _, err := metricsFetcher.getRhobsRegionAndShard()
+			if err != nil {
+				return nil, err
+			}
+
+			metricsDataSource, err := metricsFetcher.getMetricsGrafanaDataSource()
+			if err != nil {
+				return nil, err
+			}
+
+			return url.Values{
+				"var-environment":         {metricsFetcher.ocmEnvName},
+				"var-region":              {region},
+				"var-datasource_regional": {metricsDataSource},
+			}, nil
+		},
+	},
+} // Make sure to run `make generate-docs` when editing this list
+
+func GetAllowedGrafanaDashboardsShortNames() []string {
+	result := []string{}
+
+	for _, grafanaDashboard := range allowedGrafanaDashboard {
+		result = append(result, grafanaDashboard.name)
+	}
+
+	return result
+}
+
+func GetGrafanaDashboardForShortName(shortName string) *GrafanaDashboard {
+	for _, grafanaDashboard := range allowedGrafanaDashboard {
+		if grafanaDashboard.name == shortName {
+			return grafanaDashboard
+		}
+	}
+	return nil
+}
+
+func GetGrafanaDashboardUrl(metricsFetcher, logsFetcher *RhobsFetcher, grafanaDashboard *GrafanaDashboard) (string, error) {
+	dashboardParams, err := grafanaDashboard.validateAndGetParams(metricsFetcher, logsFetcher)
+	if err != nil {
+		return "", fmt.Errorf("failed to get parameters for Grafana dashboard: %v", err)
+	}
+
+	return grafanaBaseUrl + "d/" + grafanaDashboard.pathId + "/" + grafanaDashboard.pathName + "?" + dashboardParams.Encode(), nil
 }
